Upload
Code based on Google example available at: https://github.com/googlesamples/android-play-location/tree/master/LocationUpdatesForegroundService/

We have aded a login activity, a map so that we can know where are we exactly, and the crowded places according to our server.
This should work perfectly on Oreo, using a foreground service to provide the information.
We may change later to geofencing for example, so that it saves battery.